refactor: move AriaQueryHandler outside cdp
