+++ packages/puppeteer-core/src/common/GetQueryHandler.ts
@@ -4,8 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import {ARIAQueryHandler} from '../cdp/AriaQueryHandler.js';
-
+import {ARIAQueryHandler} from './AriaQueryHandler.js';
 import {CSSQueryHandler} from './CSSQueryHandler.js';
 import {customQueryHandlers} from './CustomQueryHandler.js';
 import {PierceQueryHandler} from './PierceQueryHandler.js';
